Move schema include to top of json.hpp
