+++ Test/Makefile
@@ -5,7 +5,7 @@ EFF_SRCS :=
 
 SRCS := main.cpp test_atoi.cpp test_isdigit.cpp test_memset.cpp test_strcmp.cpp \
        test_strlen.cpp test_promise.cpp test_networking.cpp test_linear_algebra.cpp test_validate_int.cpp \
-       test_task_scheduler.cpp
+       test_task_scheduler.cpp test_json_validate.cpp
 
 ifeq ($(OS),Windows_NT)
     MKDIR = mkdir
